refactor : 프로젝트 폴더 경로 수정
