+++ lib/core/data/data_source/recipe_data_source.dart
@@ -1,4 +1,5 @@
-import 'package:recipe_app/data/dto/recipe_dto.dart';
+// Project imports:
+import 'package:recipe_app/core/data/dto/recipe_dto.dart';
 
 abstract interface class RecipeDataSource {
   Future<List<RecipeDto>> fetchRecipes();
