+++ lib/core/data/dto/ingredient_dto.dart
@@ -1,3 +1,4 @@
+// Package imports:
 import 'package:json_annotation/json_annotation.dart';
 
 part 'ingredient_dto.g.dart';
