+++ lib/core/data/model/recipe.dart
@@ -1,5 +1,7 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// Project imports:
 import 'recipe_ingredient.dart';
 
 part 'recipe.freezed.dart';
