+++ lib/core/data/model/recipe_ingredient.dart
@@ -1,5 +1,7 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// Project imports:
 import 'ingredient.dart';
 
 part 'recipe_ingredient.freezed.dart';
