+++ lib/core/data/repository/recipe_repository.dart
@@ -1,0 +1,7 @@
+// Project imports:
+import 'package:recipe_app/core/data/model/recipe.dart';
+import 'package:recipe_app/core/result.dart';
+
+abstract interface class RecipeRepository {
+  Future<Result<List<Recipe>>> getRecipes();
+}
