+++ lib/core/result.dart
@@ -1,5 +1,7 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// Project imports:
 import 'failure.dart';
 
 part 'result.freezed.dart';
