+++ lib/core/ui/text_style.dart
@@ -1,4 +1,7 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Package imports:
 import 'package:google_fonts/google_fonts.dart';
 
 abstract class AppTextStyles {
