+++ lib/core/ui/ui_state.dart
@@ -1,3 +1,4 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'ui_state.freezed.dart';
