+++ lib/presentation/component/filter_search_bottom_sheet/filter_search_state.dart
@@ -1,3 +1,4 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'filter_search_state.freezed.dart';
