+++ lib/presentation/component/filter_search_bottom_sheet/rating_button.dart
@@ -1,6 +1,9 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/ui/color_style.dart';
-import 'package:recipe_app/ui/text_style.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+
+// Project imports:
 
 class RatingButton extends StatelessWidget {
   final double rate;
