+++ lib/presentation/component/image_component/skelton_animtion_widget.dart
@@ -1,5 +1,8 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/ui/color_style.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+
+// Project imports:
 
 class SkeletonAnimationWidget extends StatefulWidget {
   final double? width;
