+++ lib/presentation/component/login_component/divider_with_text.dart
@@ -1,5 +1,8 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/ui/color_style.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+
+// Project imports:
 
 class DividerWithText extends StatelessWidget {
   final String text;
