+++ lib/presentation/component/login_component/social_login_section.dart
@@ -1,4 +1,7 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Project imports:
 import 'package:recipe_app/presentation/component/login_component/divider_with_text.dart';
 import 'package:recipe_app/presentation/component/login_component/social_login_button.dart';
