+++ lib/presentation/component/rating_dialog_content.dart
@@ -1,6 +1,9 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/ui/color_style.dart';
-import 'package:recipe_app/ui/text_style.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+
+// Project imports:
 
 class RatingDialogContent extends StatefulWidget {
   final String title;
