+++ lib/presentation/component/recipe_component/recipe_card.dart
@@ -1,10 +1,11 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+import 'package:recipe_app/core/ui/ui_size.dart';
+// Project imports:
 import 'package:recipe_app/presentation/component/image_component/app_image.dart';
 import 'package:recipe_app/presentation/component/image_component/skelton_animtion_widget.dart';
-import 'package:recipe_app/ui/color_style.dart';
-import 'package:recipe_app/ui/text_style.dart';
-
-import '../../../ui/ui_size.dart';
 
 class RecipeCard extends StatefulWidget {
   final String title;
