+++ lib/presentation/component/tabs.dart
@@ -1,6 +1,9 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/ui/color_style.dart';
-import 'package:recipe_app/ui/text_style.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+
+// Project imports:
 
 class Tabs extends StatelessWidget {
   final List<String> tabTitles;
