+++ lib/presentation/saved_recipes/saved_recipes_screen.dart
@@ -1,10 +1,12 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
-import 'package:recipe_app/data/mocks/mock_recipe_data_source_impl.dart';
-import 'package:recipe_app/data/repository/recipe_repository_impl.dart';
+// Project imports:
+import 'package:recipe_app/core/data/mocks/mock_recipe_data_source_impl.dart';
+import 'package:recipe_app/core/data/repository/recipe_repository_impl.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+import 'package:recipe_app/core/ui/ui_state.dart';
 import 'package:recipe_app/presentation/component/recipe_component/recipe_card.dart';
 import 'package:recipe_app/presentation/saved_recipes/saved_recipes_view_model.dart';
-import 'package:recipe_app/ui/text_style.dart';
-import 'package:recipe_app/util/ui_state.dart';
 
 class SavedRecipesScreen extends StatelessWidget {
   final SavedRecipesViewModel viewModel;
