+++ lib/presentation/saved_recipes/saved_recipes_view_model.dart
@@ -1,8 +1,10 @@
+// Flutter imports:
 import 'package:flutter/foundation.dart';
-import 'package:recipe_app/data/model/recipe.dart';
-import 'package:recipe_app/data/repository/recipe_repository.dart';
-import 'package:recipe_app/util/result.dart';
-import 'package:recipe_app/util/ui_state.dart';
+// Project imports:
+import 'package:recipe_app/core/data/model/recipe.dart';
+import 'package:recipe_app/core/data/repository/recipe_repository.dart';
+import 'package:recipe_app/core/result.dart';
+import 'package:recipe_app/core/ui/ui_state.dart';
 
 class SavedRecipesViewModel with ChangeNotifier {
   final RecipeRepository _repository;
