+++ lib/presentation/search_recipes/search_recipes_state.dart
@@ -1,7 +1,9 @@
+// Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:recipe_app/data/model/recipe.dart';
+// Project imports:
+import 'package:recipe_app/core/data/model/recipe.dart';
+import 'package:recipe_app/core/ui/ui_state.dart';
 import 'package:recipe_app/presentation/component/filter_search_bottom_sheet/filter_search_state.dart';
-import 'package:recipe_app/util/ui_state.dart';
 
 part 'search_recipes_state.freezed.dart';
