+++ lib/presentation/splash/splash_screen.dart
@@ -1,9 +1,12 @@
+// Dart imports:
 import 'dart:ui';
 
+// Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:recipe_app/core/ui/color_style.dart';
+import 'package:recipe_app/core/ui/text_style.dart';
+// Project imports:
 import 'package:recipe_app/presentation/component/custom_button.dart';
-import 'package:recipe_app/ui/color_style.dart';
-import 'package:recipe_app/ui/text_style.dart';
 
 // ===== 애니메이션 타임라인 (시작 시점 기준) =====
 const kBackgroundStartTime = Duration(milliseconds: 0);
