+++ pubspec.yaml
@@ -55,6 +55,7 @@ dev_dependencies:
   build_runner: ^2.4.15
   freezed: ^3.0.4
   json_serializable: ^6.9.4
+  import_sorter: ^4.6.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
