+++ test/presentation/component/custom_tab_bar_test.dart
@@ -1,5 +1,10 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Package imports:
 import 'package:flutter_test/flutter_test.dart';
+
+// Project imports:
 import 'package:recipe_app/presentation/component/tabs.dart';
 
 void main() {
